Make ode solver aware of custom functions in some cases
This solves #118 but there are a few caveats. Ideally we would want to
use mathematical and external functions much more flexibly, but the
ODE solver is not really done for that and espeically when the ODEs are
solved analytically we only handle simple ODEs. Now we added the ability
to use in ODEs user functions (known C math functions and others) and
at least for numerical ODE solving, they are taken into account by the
sympy C code printer.