Fix double-free of ast nodes in python bindings
Fix construction of ast based on python object.
Example:

   >>> s = ast.String("Foo")
   >>> n = ast.Name(s)

The nmodl::ast::String instance is shared by the `s' python
object and the nmodl::ast::Name instance.

This patch ensures that the nmodl::ast::Name constructor
is given the nmodl::ast::String shared pointer, not the
raw pointer that basically would make Name also owner of the pointer,
and lead to a double free.

Fixes #161