Fix: downgrade of parent pointer to raw pointer instead of shared_ptr
shared_from_this works as intended only when the relative shared_ptr
is previously created. There are cases where this is a problem:

- in the parent constructor we cannot pass its shared_ptr to the
  children: the shared_ptr to the parent is not yet constructed
- when the parent was assigned to a raw pointer and we add a new
  child: its shared_ptr was not yet created

In addition there is the circular reference problem. The next best
option is weak_ptr. Even if this completly solves the circular
reference problem, it does not really solve the previous 2 cases.
In addition, bison can work only with raw pointers.

This solution (using a raw pointer for the parent) solves all the
previous issues and makes the code much more elegant since the parent
pointer can be updated in the "set_*", "add*" functions, and the
constructors only.