Turn every child to private
Previously, some children (i.e. statement) were left public
for easy access and manypulation by the visitors from the
parent. Thus, the interface for handling children in a safe
way (i.e. addNode) was only partially implemented. Now that
there is a parent poiner, if the paradigm was left unchanged,
every visitor would have the responsibility of updateding
children's parents. This is prone to errors. Thus, the
interface was expanded (if not completed).

- Add get_<child>_cr() to get a const ref of the child.
  This may replace completly get_<child> in the future.
- Vector iterface members that closely resemble std::vector
  - add<child> is the corrispective of vector::emplace_back
    It does not have the same name due to legacy code. It
    can be changed with a little bit of refactoring
  - insert<child>
  - erase<child>
  - reset<child> resets the shared_ptr
- Mark clone() as const
- Mark eval() as const
- Adapt visitors to the new interface. They use only the
  new _cr getters
- Add ckParent_visitor to most of the visitor tests (at
  the end, after the visitor passed) to check that
  parents are still up to date.