Remove the jinja macro "override(node)"
override(node) is a macro to add override only where needed and
copied from virtual(node). However, override is required also
for derived classes were virtual is not needed (we do not want
a virtual table). It was used only for get_token and set_name,
2 methods that are in ast_commons and, thus, must be always
overridden. For these reasons override is actually superflous
and does not print "override" for LinearBlock where it should
(because LinearBlock is a virtual class still). Removing it
and hardcoding override in the 2 functions erases the problem
and simplifies the code.