Bugfix: changed cmake to pass CI for fmt libs used in testparser
The problem is presenting only when building in release version,
when checkparent_visitor is used in testparser, and when fmt libs
are used in checkparent_visitor. However, the solution is not to
link the fmt libraries in the cmakelist. The solution is to target
link the libraries printer and symtab for testparser. Since fmt
should be header only and visitor (in which there is
checkparent_visitor) is static, I do not see why this problem
appears. And why does it appear only for the release version?
The debug version should contain more linking problems (usually).
FOr now this is the fix. Still not sure of the cause.