Add support for for_netcon (issue #390)
For_netcon applies the operations of the block to the weights of
all the netcons. Since all the weights are on the same vector,
weights, this means appling the same operations to a mask of values
with the appropriate offset. The mask is the arguments of the
for_netcon block.

Note: the local variables are renamed to the correct position
in the corenrn vectors (no new variables added). Some of these
local variables may have been already substituted with special
characters that break rename_visitor (i.e. (*w) ). Thus, we need
to sanitize first the regex names that we are going to substitute.

- Tested with the test in /nrn/test/coreneuron/test_fornetcon.py
for SoA

TODO: Test for AoS
TODO: Add test to a CI