Make tmp* variables unique
We leverage the fact that suffix_random_string now returns
a prefix that is unique to create various unique strings in
ode.py appending the usual iterator. The tmp_unique_prefix
is provided to solve_lin_system. Only in case do_cse is True
we use the tmp_unique_prefix to generate the temp variables.

- remove _make_unique_prefix, it was not useful and code was
  duplicated
- tmp_* is now the default to better distinguish the iterator