IR builder redesign for LLVM IR code generation pipeline (#634)
Improves the code structure for the LLVM code generation pipeline

The following changes were added:

1. New IR builder class.

Before, LLVM visitor just simply used `llvm::IRBuilder<>` class to
generate instructions. Recently, this (as well as adding the functionality
to the visitor on the go) had led to code duplication and it became hard
to introduce new features nicely.

Hence, a special `IRBuilder` class is now used. This class is a wrapper
around `llvm::IRBuilder<>` that keeps track of certain IR generation
specific fields (that are unrelated to the visitor), defines an API that the
visitor can use to generate LLVM IR.

Also, this IR builder has been designed to be nearly fully-independent from
NMODL AST nodes. this allows it to be more generic and to be more
extensible.

2. Visitor clean-up

 LLVM visitor has been refactored to take the new IR builder class into
account. Also, the functions were reordered, refactored and renamed to
better reflect the intended use and provide encapsulation.

3. Scatter preparation

The functionality of the generating code for `CodegenInstanceVar` node
has been extended with `read_from_or_write_to_instance(...)` function.
Now, an optional `value_to_store` is passed to indicate whether the code
needs to be generated for reading the instance variable or writing to it.


fixes #538