Added attributes and metadata to LLVM IR compute kernels (#648)
Previously, there was no metadata and attributes associated with the
instance struct pointer, compute kernels or loops. This commit fixes
this.

- New instance struct attributes

Since all pointers contained in the instance struct do not alias, we add
a `noalias` (LLVM's `__restrict` alternative) attribute to it. In addition,
we add `nocapture` (No capturing occurs in the function) and
`readonly` (Struct pointer is not written to) attributes.

This means that some load instructions can be moved out from the loop
body. Example:
```llvm
; BEFORE
for.body.lr.ph:                                   ; preds = %0
  %5 = getelementptr inbounds %avx__instance_var__type, %avx__instance_var__type* %mech1, i64 0, i32 1
  br label %for.body

for.body:                                         ; preds = %for.body.lr.ph, %for.body
  %15 = load double*, double** %5, align 8
  ; ...


; AFTER
for.body.lr.ph:                                   ; preds = %0
  %5 = getelementptr inbounds %avx__instance_var__type, %avx__instance_var__type* %mech1, i64 0, i32 1
  %6 = load double*, double** %5, align 8
  br label %for.body
```

- New function attributes

Now, compute kernels are marked with `nofree` and `nounwind`
attributes.

- Loop metadata

Also, loop metadata is added to scalar kernels, specifying that no
vectorization is needed. The reason for this is because we want to
benchmark truly scalar kernels, and disable LLVM's vectorization if
necessary.

Note that for vector loop epilogue there is no metadata that disables
vectorization.

fixes #607