Basic predication support for LLVM backend (#652)
Added support for vector predication. Currently, we support a very basic
predication pattern (that will be extended in the future):
```c++
IF (/*condition*/) {
  // code here, no nested conditionals
} ELSE {
  // code here, no nested conditionals
} 
```
**What has been changed and added**

1. Removed vectorization check

Before, in the `FOR` statement visitor we were checking whether the code
can be vectorized. After refactoring `llvm::IRBuilder<>` into a separate class,
there is no interface to reset the builder's vector width. Hence, this check leads
to visitor having scalar vector width of 1, and builder having the same vector width.
```c++
if (!can_vectorize(node, sym_tab)) {
    vector_width = 1;
     ir_builder.generate_scalar_code();
}
```
In order to avoid any issues, this check is simply removed and will be added in
the separate PR.

2. Predication support

- `can_vectorize` has been changed to support a single `IF` or `IF/ELSE` pair.
- A special vectorized `IF` AST node visitor has been added.
- If generating code within `IF` AST node, instructions are masked.

3. Added execution and IR tests

fixes #539