[CP-870] Cherry-pick sympy fix from master (#872)
* Squash commit when rebasing with master

* [CP-870] fixup! CI errors with sympy 1.9 and 1.10 (#870)

* fixup! CI errors with sympy 1.9 and 1.10

- bin is a function in python: renamed variable in test to not get this error
- sympy now can solve z'=a/z+b/z/z. It is getting hard to find simple equations
that it cannot solve.

Warning: this is a downgrade of the testing capabilities. I could not find
an ode with simple functions (no sin, exp, log etc. Only simple multiplications
and additions) that cannot be solved by sympy. Best case the test hangs waiting
for sympy to solve a very long equation. For this reason, with this PR, the
fact that the code should return the equation untouched if it cannot be processed
is untested

* fixup! clang-format

* remove limit for sympy 1.9

* Relax constraints of GPU nodes to allocate in gitlab CI

Co-authored-by: Alessandro Cattabiani <cattabiani@users.noreply.github.com>
Co-authored-by: Alessandro Cattabiani <alessandro.cattabiani@epfl.ch>