Make `generate_references.py` configurable.
Adds the ability to customize with which flags `nmodl` is run to
generate the references.

The default is `""` (for CoreNEURON) and `"--neuron"` (for NEURON).
These choices can be overridden by adding a file `references.json`
to the usecase directory.

Since the logic is slightly more involved, the Bash script is
reimplemented as a Python script.