Include solvers to avoid include path dependency. (#1329)
Similar to #1305, include the solver headers in variables in a C++
header, and include them in the code generated for the MOD files. This
should allow to substitute different NMODL binaries in NEURON builds
without confusion as to which solver headers are going to be included in
the MOD file compilation.