Homogenize finding our components.
We currently rely on two files being present in pre-defined directories:

* nrnunits.lib
* libpywrapper.so (or similar)

These are searched for in:

* CMake's binary directory for NMODL
* CMake's install directory for NMODL
* The environment variable `NMODLHOME`

Leading to some acrobatics to always set NMODLHOME correctly.  This PR
attempts to also find the above files relative to the NMODL executable,
in the hope that some use cases of NMODLHOME can be removed.