Remove `inline` from function definitions.
Since the compiler decides to inline regardless of the keyword `inline`,
it's not helpful for performance reasons.

The issue is that if we accidentally define multiple, incompatible
copies of the same symbol, when linking we'll got get an error, instead
the linker will pick one of the definitions at random.

Since the compiler, might decide to inline, we might not immediately
notice the bug, e.g. through integration tests.