Get voltage via `prop->node`.
When calling functions directly, don't have access to the `NrnThread`
and therefore can't get node properties, like the voltage from there.

The solution is to create a link from the Prop to the Node. Then use
that link to figure out the node properties.

The trick to make the two cases uniform is the same as we use for
instance data, we create pointers to array of length one (by taking the
address of the element) and setting `_iml`/`id` to `0`.