Print tracebacks of Python exceptions.
The SymPy solvers are meant to be run again on code that has already
been transformed. However, since the output of the SymPy solvers isn't
valid input for the SymPy solver they crash the second time around.

Failure is interpreted as: "it already ran" and not considered a
failure.

Naturally, this is rarely the case while developing and it's always a
bug.

The workaround here is to print the traceback as a separate log message
of severity "info". Unfortunately, `pybind11` injects a line:

    # -*- coding: utf-8 -*-

Hence, all line numbers are off by two; because `ode_py` starts with an
empty line.

Old output:

    [NMODL] [warning] :: SympySolverVisitor :: solve_non_lin_system python exception: name 'x' is not defined

New output:

    [NMODL] [warning] :: SympySolverVisitor :: solve_non_lin_system python exception. (--verbose=info)
    [NMODL] [info] :: Traceback (most recent call last):
      File "<string>", line 691, in <module>
      File "<string>", line 416, in solve_non_lin_system
      File "<string>", line 385, in finite_difference_variables
      File "<string>", line 382, in recurse
      File "<string>", line 352, in finite_difference_step_variable
    NameError: name 'x' is not defined