Print more information on failure. (#1441)
Currently, on failure it will output:

```
terminate called after throwing an instance of 'std::runtime_error'
  what():  NMODL Parser Error : syntax error, unexpected / [Location : 1.40]
---------------------------------------^
```

but it's entirely unclear what it was processing when it crashed. It's
also not obvious from the surrounding context when building NEURON.

This PR changes it such that we have:
```
[FATAL] NMODL encountered an unhandled exception.
    cwd = "build/test/nrnivmodl/53c4"
    nmodl Aradi_CadepK.mod --neuron -o build/test/nrnivmodl/53c4/x86_64
```
followed by the exception.