Use compiled version of spdlog. (#1476)
We see that in NMODL using the header-only version of
SPDLOG contributes noticeably to the compile time:
```
**** Template sets that took longest to instantiate:
 52766 ms: spdlog::pattern_formatter::handle_flag_<$> (92 times, avg 573 ms)
```
However, this can be easily avoided by using a non-headeronly
version of SPDLOG.

When using an external copy of spdlog it'll pick up the shared
library. When built from the submodule it creates a static library.